+++ packages/tooltip-vue/vitest.config.ts
@@ -1,7 +1,10 @@
-import { defineConfig } from 'vitest/config'
+import { createWorkspaceVitestConfig } from "../../config/vitest.base"
 
-export default defineConfig({
+export default createWorkspaceVitestConfig(import.meta.url, {
   test: {
-    passWithNoTests: true,
+    environment: "jsdom",
+    coverage: {
+      include: ["src/**/*.ts"],
+    },
   },
 })